Events: refactoring
