+++ src/js/components/events/list/line/index.d.tsx
@@ -1,0 +1,13 @@
+import { IEvents } from '~store/events/reducer'
+export { IStore } from '~store/index'
+
+export interface IOwnProps {
+  eventId: IEvents.IEvent['eventId']
+  eventsPlaces: IEvents.IState['places']
+  handleSelected
+}
+
+export interface IReduxProps {
+  eventData: IEvents.IEvent
+  isSelected: boolean
+}
