+++ src/js/components/events/list/popups/addEvent/index.tsx
@@ -1,3 +1,4 @@
+import * as I from './index.d'
 import React from 'react'
 import { connect } from 'react-redux'
 import produce from 'immer'
@@ -7,46 +8,18 @@ import DatePicker from 'react-day-picker/DayPickerInput'
 import DatePickerLocalizaton, { formatDate, parseDate } from 'react-day-picker/moment'
 import * as eventsSelectors from '~store/events/selectors'
 import * as eventsActions from '~store/events/actions'
-import styles from './styles.pcss'
-import { IStore } from '~store/index'
+import styles from '../styles.pcss'
 import 'moment/locale/ru'
-
-interface IPopupProps {
-  isOpen: boolean | null
-  handleOpen: any
-}
 
 PopupAddEvent.portalId = `popupAddEvent-${Math.random()}`
 PopupAddEvent.classList = [styles.popupAddEvent]
 
-function PopupAddEvent(props: IPopupProps) {
+function PopupAddEvent(props: I.IPopupProps) {
   return (
     <PopupPortal isOpen={props.isOpen} portalId={PopupAddEvent.portalId} classList={PopupAddEvent.classList}>
       <FormAddEvent handleOpen={props.handleOpen} />
     </PopupPortal>
   )
-}
-
-interface IReduxProps {
-  eventsPlaces: IStore['events']['places']
-  isSaving: IStore['events']['isSaving']
-}
-
-interface IDispatchProps {
-  saveEvent: any
-}
-
-interface IOwnProps {
-  handleOpen: IPopupProps['handleOpen']
-}
-
-type IProps = IOwnProps & IReduxProps & IDispatchProps
-
-interface IState {
-  fieldEventName: string
-  fieldEventDate: string
-  fieldEventPlace: string
-  errors: { [errorName: string]: Error }
 }
 
 enum fieldNames {
@@ -57,14 +30,14 @@ enum fieldNames {
 
 const dateFormat = 'DD.MM.YYYY'
 
-class Form extends React.Component<IProps, IState> {
+class Form extends React.Component<I.IProps, I.IState> {
   defaultState = { fieldEventName: '', fieldEventDate: '', fieldEventPlace: '', errors: {} }
   state = this.defaultState
 
   componentDidCatch(error: Error) {
     this.handleError({ errorName: `componentDidCatch:${error.message}`, error })
   }
-  setStateProxy = (func: ((state: IState) => void)) => this.setState(produce(func))
+  setStateProxy = (func: (state: I.IState) => void) => this.setState(produce(func))
 
   readonly handleError = ({ errorName, error }: { errorName: string; error: Error }) =>
     this.setState({ errors: { ...this.state.errors, [errorName]: error } })
@@ -157,15 +130,12 @@ class Form extends React.Component<IProps, IState> {
   }
 }
 
-const mapStateToProps = (store: IStore): IReduxProps => ({
+const FormAddEvent = connect(
+  (store: I.IStore): I.IReduxProps => ({
     eventsPlaces: eventsSelectors.placesGetter(store),
     isSaving: eventsSelectors.isSaving(store),
   }),
-  mapDispatchToProps = { saveEvent: eventsActions.saveEvent }
-
-const FormAddEvent = connect(
-  mapStateToProps,
-  mapDispatchToProps
+  { saveEvent: eventsActions.saveEvent } as I.IDispatchProps
 )(Form)
 
 export default PopupAddEvent
