+++ src/js/components/events/list/popups/deleteEvents/index.tsx
@@ -1,27 +1,12 @@
+import * as I from './index.d'
 import React from 'react'
 import { connect } from 'react-redux'
 import { PopupPortal } from '~modules/popup'
 import * as eventsActions from '~store/events/actions'
 import * as eventsSelectors from '~store/events/selectors'
-import { IStore } from '~store/index'
-import styles from './styles.pcss'
+import styles from '../styles.pcss'
 
-interface IReduxProps {
-  isDeleting: IStore['events']['isDeleting']
-}
-
-interface IDispatchProps {
-  deleteEvents: any
-}
-
-interface IOwnProps {
-  isOpen: boolean | null
-  handleOpen: any
-}
-
-type IProps = IOwnProps & IReduxProps & IDispatchProps
-
-class FormDeleteEvents extends React.Component<IProps> {
+class FormDeleteEvents extends React.Component<I.IProps> {
   static portalId = `popupDeleteEvents-${Math.random()}`
   static classList = [styles.popupDeleteEvents]
 
@@ -65,8 +50,8 @@ class FormDeleteEvents extends React.Component<IProps> {
 }
 
 const PopupDeleteEvents = connect(
-  store => ({ isDeleting: eventsSelectors.isDeleting(store) }),
-  { deleteEvents: eventsActions.deleteEvents }
+  (store: I.IStore): I.IReduxProps => ({ isDeleting: eventsSelectors.isDeleting(store) }),
+  { deleteEvents: eventsActions.deleteEvents } as I.IDispatchProps
 )(FormDeleteEvents)
 
 export default PopupDeleteEvents
