+++ src/js/store/events/reducer.d.ts
@@ -40,3 +40,28 @@ export interface IState {
     searchTerm: string
   }
 }
+
+export enum ActionTypes {
+  EVENTS_FETCH = 'EVENTS_FETCH',
+  EVENTS_FETCH_REQUEST = 'EVENTS_FETCH_REQUEST',
+  EVENTS_FETCH_SUCCESS = 'EVENTS_FETCH_SUCCESS',
+  EVENTS_FETCH_FAIL = 'EVENTS_FETCH_FAIL',
+
+  EVENT_SAVE = 'EVENT_SAVE',
+  EVENT_SAVE_REQUEST = 'EVENT_SAVE_REQUEST',
+  EVENT_SAVE_SUCCESS = 'EVENT_SAVE_SUCCESS',
+  EVENT_SAVE_FAIL = 'EVENT_SAVE_FAIL',
+
+  EVENTS_DELETE = 'EVENTS_DELETE',
+  EVENTS_DELETE_REQUEST = 'EVENTS_DELETE_REQUEST',
+  EVENTS_DELETE_SUCCESS = 'EVENTS_DELETE_SUCCESS',
+  EVENTS_DELETE_FAIL = 'EVENTS_DELETE_FAIL',
+
+  EVENT_SELECTED = 'EVENT_SELECTED',
+  EVENTS_SEARCH = 'EVENTS_SEARCH',
+
+  PLACES_FETCH = 'PLACES_FETCH',
+  PLACES_FETCH_REQUEST = 'PLACES_FETCH_REQUEST',
+  PLACES_FETCH_SUCCESS = 'PLACES_FETCH_SUCCESS',
+  PLACES_FETCH_FAIL = 'PLACES_FETCH_FAIL',
+}
