+++ tsconfig.json
@@ -28,7 +28,7 @@
     "strictPropertyInitialization": true,
     "strictNullChecks": true,
     "importHelpers": true,
-    "isolatedModules": true,
+    "isolatedModules": false,
     "baseUrl": "./",
     "paths": {
       "~components/*": ["src/js/components/*"],
